requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.34.23
botocore==1.34.46
build==1.1.1
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.9
django-cors-headers==4.3.1
django-quill-editor==0.1.40
django-storages==1.14.2
djangorestframework==3.14.0
docutils==0.20.1
drf-extra-fields==3.7.0
filelock==3.13.1
filetype==1.2.0
gunicorn==21.2.0
idna==3.6
importlib_metadata==7.0.2
jaraco.classes==3.3.1
jmespath==1.0.1
keyring==24.3.1
kombu==5.3.7
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.2.0
nh3==0.2.15
packaging==23.2
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pyee==11.1.0
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-ffmpeg==2.0.10
pytz==2023.3.post1
readme_renderer==43.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
twine==5.0.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
zipp==3.18.1
